fastapi/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
av==11.0.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cryptography==42.0.5
ctranslate2==4.0.0
decorator==4.4.2
distro==1.9.0
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
fastapi==0.110.0
faster-whisper==1.0.1
filelock==3.13.1
flatbuffers==24.3.7
frozenlist==1.4.1
fsspec==2024.2.0
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
humanfriendly==10.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
instaloader==4.10.3
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
numpy==1.26.4
onnxruntime==1.17.1
openai==0.28.0
opencv-python==4.9.0.80
packaging==23.2
passlib==1.7.4
pathlib==1.0.1
pillow==10.2.0
proglog==0.1.10
protobuf==4.25.3
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
pyreadline3==3.4.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytube==15.0.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rsa==4.9
setuptools==69.1.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
stripe==8.8.0
sympy==1.12
tiktoken==0.6.0
tokenizers==0.15.2
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.28.0
watchfiles==0.21.0
wheel==0.43.0
yarl==1.9.4
